IGAPP-636: Tidy up
