+++ native/src/hooks/useLoadCityContent.ts
@@ -1,4 +1,3 @@
-import moment from 'moment'
 import { useCallback } from 'react'
 
 import {
@@ -8,21 +7,18 @@ import {
   createEventsEndpoint,
   createLanguagesEndpoint,
   createPOIsEndpoint,
-  Endpoint,
   ErrorCode,
   EventModel,
-  fromError,
   LanguageModel,
   PoiModel,
   ReturnType,
   useLoadAsync,
 } from 'api-client'
 
-import { SnackbarType } from '../components/SnackbarContainer'
 import { LanguageResourceCacheStateType } from '../redux/StateType'
 import { dataContainer } from '../utils/DefaultDataContainer'
-import { determineApiUrl } from '../utils/helpers'
 import useLoadCities from './useLoadCities'
+import useLoadWithCache from './useLoadWithCache'
 import useOnLanguageChange from './useOnLanguageChange'
 import useSnackbar from './useSnackbar'
 
@@ -41,62 +37,13 @@ export type CityContentData = {
   pois: PoiModel[]
   resourceCache: LanguageResourceCacheStateType
 }
+
 export type CityContentReturn = Omit<ReturnType<CityContentData>, 'error'> & { error: ErrorCode | Error | null }
 
-type Load<T> = {
-  cityCode: string
-  languageCode: string
-  createEndpoint: (baseUrl: string) => Endpoint<{ city: string; language: string }, T>
-  isAvailable: (cityCode: string, languageCode: string) => Promise<boolean>
-  getFromDataContainer: (cityCode: string, languageCode: string) => Promise<T>
-  setToDataContainer: (cityCode: string, languageCode: string, data: T) => Promise<void>
-  forceUpdate?: boolean
-  showSnackbar: (snackbar: SnackbarType) => void
-}
-
-const loadWithCache = async <T>({
-  cityCode,
-  languageCode,
-  isAvailable,
-  getFromDataContainer,
-  setToDataContainer,
-  createEndpoint,
-  showSnackbar,
-  forceUpdate = false,
-}: Load<T>): Promise<T | null> => {
-  const cachedData = (await isAvailable(cityCode, languageCode))
-    ? await getFromDataContainer(cityCode, languageCode)
-    : null
-
-  const lastUpdate = await dataContainer.getLastUpdate(cityCode, languageCode)
-  const shouldUpdate = forceUpdate || !lastUpdate || lastUpdate.isBefore(moment.utc().startOf('day'))
-
-  if (!shouldUpdate && cachedData) {
-    return cachedData
-  }
-
-  try {
-    const payload = await createEndpoint(await determineApiUrl()).request({
-      city: cityCode,
-      language: languageCode,
-    })
-    if (payload.data) {
-      await setToDataContainer(cityCode, languageCode, payload.data)
-    }
-    return payload.data ?? cachedData
-  } catch (e) {
-    if (cachedData) {
-      showSnackbar({ text: fromError(e) })
-    } else {
-      throw e
-    }
-  }
-  return cachedData
-}
-
-const useLoadWithCache = <T>(params: Load<T>) =>
-  useLoadAsync(useCallback(forceUpdate => loadWithCache({ ...params, forceUpdate }), [params]))
-
+/**
+ * Hook to load all the offline available city content at once and handle errors, loading and refreshing at the same time.
+ * Takes care of updating the data regularly.
+ */
 const useLoadCityContent = ({ cityCode, languageCode }: Params): CityContentReturn => {
   const citiesReturn = useLoadCities()
   const previousLanguageCode = useOnLanguageChange({ languageCode })
@@ -132,12 +79,11 @@ const useLoadCityContent = ({ cityCode, languageCode }: Params): CityContentRetu
     setToDataContainer: dataContainer.setPois,
   })
 
-  // TODO IGAPP-636: Actually load resource cache
+  // TODO IGAPP-636: Actually load resource cache and move to separate hook
   const loadResourceCache = useCallback(
     async () => dataContainer.getResourceCache(cityCode, languageCode),
     [cityCode, languageCode]
   )
-
   const resourceCacheReturn = useLoadAsync(loadResourceCache)
 
   const city = citiesReturn.data?.find(it => it.code === cityCode)
@@ -172,6 +118,7 @@ const useLoadCityContent = ({ cityCode, languageCode }: Params): CityContentRetu
     eventsReturn.loading ||
     poisReturn.loading ||
     resourceCacheReturn.loading
+
   const refresh = () => {
     citiesReturn.refresh()
     languagesReturn.refresh()
